Second version of the music map
Is now obsolete but uploading just to keep a safe copy. This version is upside down: the subject is the people looking for musicians but there is not musicians data base. Therefore a new concept is going to be done where there is both a musician table and an offers table.